requirements.txt
click==6.7
cycler==0.10.0
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.2
gpiozero==1.4.0
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
#matplotlib==2.0.2
numpy==1.13.1
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.6.1
python-engineio==1.7.0
python-socketio==1.8.0
pytz==2017.2
RPi.GPIO==0.6.3
six==1.10.0
SQLAlchemy==1.1.13
virtualenv==15.0.3
Werkzeug==0.12.2
WTForms==2.1
